requirements.txt
BTrees==4.5.1
Click==7.0
Flask==1.0.2
Flask-SocketIO==3.3.2
itsdangerous==1.1.0
Jinja2==2.10
kafka-python==1.4.4
MarkupSafe==1.1.1
numpy==1.16.2
opencv-python==4.0.0.21
persistent==4.4.3
pync==2.0.3
python-dateutil==2.8.0
python-engineio==3.4.3
python-socketio==3.1.2
six==1.12.0
transaction==2.4.0
Werkzeug==0.14.1
zc.lockfile==1.4
